Use more generic wildcard for kernel package removal
We need remove the eventually installed KVM kernel package as it currently doesn't include the binder and ashmem kernel modules. Installing the generic kernel will bring both and is what we need here. Using a wildcard and removing any installed kernel is best as not all images multipass may will come with the KVM kernel and we can ensure we end up with the generic kernel in any case.

Co-authored-by: Michał Sawicz <michal@sawicz.net>